requirements.txt
tqdm
requests
argparse
torch
matplotlib
python-dotenv
